+++ js/main.js
@@ -1,6 +1,6 @@
 import {formChangeOnUserInput} from './form-change-on-user-input.js';
 import {formDisable, mapFiltersDisable, formValidation} from './form.js';
-import {getAdvertisments} from './get-server-data.js';
+import {getAdvertisments} from './processing-server-data.js';
 import {loadMap} from './map.js';
 
 //const ADVERTISEMENT_COUNT = 10;
